Refactoring
8. String to Integer (atoi)
https://leetcode.com/problems/string-to-integer-atoi/
