+++ Math/8. String to Integer (atoi)/string-to-integer-atoi.playground/contents.xcplayground
